dontaion page
An Experimental donation page is added, it has a donationbar and donationbody elements. When user cames to donation page he/she is met with an infromative text in donationbody if the user presses the shocard in donationbar, donationbody's ingredients will be changed with donation links(Mama Kumbarası). Page does what required bu will need style update.